interfaces: use PLUGIN_DIR instead of SYSREPO_DIR for interface descrption file storage
SYSREPO_DIR will only be set when setup-dev scripts are used.
Another issue is that to use the interface description file store,
the user has to manually create the whole subpath where the file was
expected.

Instead, this expects just PLUGIN_DIR, and automatically creates a file
for description storage if it doesn't exist at plugin startup.

Signed-off-by: Juraj Vijtiuk <juraj.vijtiuk@sartura.hr>